refactor: split agent registry
