+++ alpha_factory_v1/backend/agents/climate_risk_agent.py
@@ -117,7 +117,7 @@ except Exception:  # pragma: no cover - optional
 # Alpha‑Factory locals (NO heavy deps)
 # ────────────────────────────────────────────────────────────────────────────────
 from backend.agent_base import AgentBase  # pylint: disable=import-error
-from backend.agents import AgentMetadata, register_agent
+from backend.agents.registry import AgentMetadata, register_agent
 from backend.orchestrator import _publish  # re‑use event bus
 from alpha_factory_v1.utils.env import _env_int
