+++ alpha_factory_v1/backend/agents/cyber_threat_agent.py
@@ -116,7 +116,7 @@ except ModuleNotFoundError:  # pragma: no cover
 # Alpha‑Factory locals (no heavy deps)
 # ---------------------------------------------------------------------------
 from backend.agent_base import AgentBase  # pylint: disable=import‑error
-from backend.agents import AgentMetadata, register_agent
+from backend.agents.registry import AgentMetadata, register_agent
 from backend.orchestrator import _publish  # reuse orchestrator event bus
 from alpha_factory_v1.utils.env import _env_int
