+++ alpha_factory_v1/backend/agents/retail_demand_agent.py
@@ -105,7 +105,7 @@ except Exception:  # pragma: no cover - optional
 # Alpha‑Factory lightweight core imports  ------------------------------------
 # ---------------------------------------------------------------------------
 from backend.agent_base import AgentBase  # pylint: disable=import-error
-from backend.agents import AgentMetadata, register_agent
+from backend.agents.registry import AgentMetadata, register_agent
 from backend.orchestrator import _publish  # structured‑event helper
 from alpha_factory_v1.utils.env import _env_int
