+++ alpha_factory_v1/backend/agents/smart_contract_agent.py
@@ -117,7 +117,7 @@ except ModuleNotFoundError:  # pragma: no cover
 # Alpha‑Factory local imports (never heavy)
 # ---------------------------------------------------------------------------
 from backend.agent_base import AgentBase  # pylint: disable=import-error
-from backend.agents import AgentMetadata, register_agent
+from backend.agents.registry import AgentMetadata, register_agent
 from backend.orchestrator import _publish
 from alpha_factory_v1.utils.env import _env_int
