+++ alpha_factory_v1/backend/agents/supply_chain_agent.py
@@ -115,7 +115,7 @@ except Exception:  # pragma: no cover - optional
 # Alpha‑Factory local imports (lightweight, no heavy deps)
 # ---------------------------------------------------------------------------
 from backend.agent_base import AgentBase  # pylint: disable=import‑error
-from backend.agents import AgentMetadata, register_agent  # pylint: disable=import‑error
+from backend.agents.registry import AgentMetadata, register_agent  # pylint: disable=import‑error
 from backend.orchestrator import _publish  # re‑use event bus hook
 
 logger = logging.getLogger(__name__)
