+++ tests/test_agent_manager_consumer.py
@@ -41,9 +41,9 @@ def test_manager_starts_and_stops_bus_consumer(monkeypatch: pytest.MonkeyPatch) 
         pass
 
     monkeypatch.setattr("alpha_factory_v1.backend.agent_manager.EventBus", DummyBus)
-    monkeypatch.setattr("backend.agents.list_agents", list_agents)
-    monkeypatch.setattr("backend.agents.get_agent", get_agent)
-    monkeypatch.setattr("backend.agents.start_background_tasks", start_background_tasks)
+    monkeypatch.setattr("backend.agents.registry.list_agents", list_agents)
+    monkeypatch.setattr("backend.agents.registry.get_agent", get_agent)
+    monkeypatch.setattr("backend.agents.health.start_background_tasks", start_background_tasks)
     monkeypatch.setattr("alpha_factory_v1.backend.agent_runner.get_agent", get_agent)
 
     mgr = AgentManager({"dummy"}, True, None, 60, 30)
